% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McMV_test.R
\name{McMV_test}
\alias{McMV_test}
\title{mean-variance-based method:McMV,simultaneously test for differences in both the means and variances of a quantitative trait.}
\usage{
McMV_test(
  Genotype,
  Y,
  Sex,
  Covariate = NULL,
  missing_cutoff = 0.15,
  MAF_Cutoff = NULL,
  MGC_Cutoff = 20,
  kins = NULL,
  method = "joint"
)
}
\arguments{
\item{Genotype}{A numeric genotype matrix with each row as a different individual and each column as a separate SNP.
Each genotype is coded as 0, 1 or 2 for females and coded as 0 or 1 for males, indicating the number of reference allele.}

\item{Y}{A numeric vector of a quantitative trait, such as height.}

\item{Sex}{A vector of the genetic sex following PLINK default coding, where males are coded as 1 and females are coded as 2.}

\item{Covariate}{ptional: a vector or a matrix of covariates, such as age.}

\item{missing_cutoff}{Cutoff of the missing rates of SNPs (default=0.15). Any SNPs with missing rates higher than the cutoff will be excluded from the analysis.}

\item{MAF_Cutoff}{MAF cutoff for common vs. rare variants (default=NULL). It should be a numeric value between 0 and 0.5, or NULL.}

\item{MGC_Cutoff}{Cutoff for the minimum genotype count in either females or males.}

\item{kins}{For the mixed data, kins=GRM is a block matrix, interpreted as the genetic relatedness among these individuals,
including general pedigrees and unrelated individuals.
For general pedigrees, GRM is a kinship matrix, indicating the genetic relatedness among individuals within general pedigrees,
calculated using the dedicated method for kinship coefficients of X chromosome provided by the “kinship2” package in R software.}

\item{method}{Optional: A character string indicating which kind of association tests is to be conducted.
There are four options: "MTchenw", "variance", "joint" (default) and "all".
method="MTchenw": MTchenw; method="variance": MwM3VNA; method="joint": McMV; method="all": all of the above association tests.}
}
\value{
The p values and the test statistics of association tests selected by the method option for each SNP.
}
\description{
A function to obtain the p value and the test statistics of the MTchenw_test (i.e., MTchenw), the variance test (i.e.,MwM3VNA), the McMV_test(i.e., McMV) or all.
}
\examples{
McMV_test(Genotype,Y,Sex,
                    Covariate=mixed[,"Age"],
                    missing_cutoff=0.15,
                    MAF_Cutoff=NULL,
                    MGC_Cutoff=20,
                    kins=GRM,
                    method='joint')
}
