% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTchenw_test.R
\name{MTchenw_test}
\alias{MTchenw_test}
\title{mean-based method: MTchenw, including a dominant effect of heterozygous (Aa) females}
\usage{
MTchenw_test(
  Genotype,
  Y,
  Sex,
  Covariate = NULL,
  missing_cutoff = 0.15,
  MAF_Cutoff = NULL,
  MGC_Cutoff = 20,
  kins = kins
)
}
\arguments{
\item{Genotype}{A numeric genotype matrix with each row as a different individual and each column as a separate SNP.
Each genotype is coded as 0, 1 or 2 for females and coded as 0 or 1 for males, indicating the number of reference allele.}

\item{Y}{A numeric vector of a quantitative trait, such as height.}

\item{Sex}{A vector of the genetic sex following PLINK default coding, where males are coded as 1 and females are coded as 2.}

\item{Covariate}{Optional: a vector or a matrix of covariates, such as age.}

\item{missing_cutoff}{Cutoff of the missing rates of SNPs (default=0.15).
Any SNPs with missing rates higher than the cutoff will be excluded from the analysis.}

\item{MAF_Cutoff}{MAF cutoff for common vs. rare variants (default=NULL). It should be a numeric value between 0 and 0.5, or NULL.
When it is NULL, 1/ sqrt(2 SampleSize) will be used (Ionita-Laza et al. 2013). Only common variants are included in the analysis.}

\item{MGC_Cutoff}{Cutoff for the minimum genotype count in either females or males.}

\item{kins}{For the mixed data, kins is a block matrix, interpreted as the genetic relatedness among these individuals,
including general pedigrees and unrelated individuals.
For general pedigrees, it is a kinship matrix, indicating the genetic relatedness among individuals within general pedigrees,
calculated using the dedicated method for kinship coefficients of X chromosome provided by the “kinship2” package in R software.}
}
\value{
The p value and test statistic of MTchenw.
}
\description{
A function to obtain the p value and the test statistic of MTchenw for testing the mean difference of the trait value across genotypes.
}
