% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tplinkw_test.R
\name{Tplinkw_test}
\alias{Tplinkw_test}
\title{mean-based method: Tplinkw for unrelated data}
\usage{
Tplinkw_test(
  Genotype,
  Y,
  Sex,
  Covariate = NULL,
  missing_cutoff = 0.15,
  MAF_Cutoff = NULL,
  MGC_Cutoff = 20
)
}
\arguments{
\item{Genotype}{A numeric genotype matrix with each row as a different individual and each column as a separate SNP.
Each genotype is coded as 0, 1 or 2 for females and coded as 0 or 1 for males, indicating the number of reference allele.}

\item{Y}{A numeric vector of a quantitative trait, such as height.}

\item{Sex}{A vector of the genetic sex following PLINK default coding, where males are coded as 1 and females are coded as 2.}

\item{Covariate}{ptional: a vector or a matrix of covariates, such as age.}

\item{missing_cutoff}{Cutoff of the missing rates of SNPs (default=0.15). Any SNPs with missing rates higher than the cutoff will be excluded from the analysis.}

\item{MAF_Cutoff}{MAF cutoff for common vs. rare variants (default=NULL). It should be a numeric value between 0 and 0.5, or NULL.}

\item{MGC_Cutoff}{Cutoff for the minimum genotype count in either females or males.}
}
\value{
The p value and test statistic of Tplinknw.
}
\description{
#' Similar to the MTplink method, Tplinkw is an approach designed for unrelated individuals and
therefore does not incorporate a kinship matrix to account for correlations between individuals.
}
\examples{
Tplinkw_test(Genotype,Y,Sex,
                       Covariate=unrelated[,"Age"],
                       missing_cutoff=0.15,
                       MAF_Cutoff=NULL,
                       MGC_Cutoff=20)
}
